Student fees

+++ app/course/course.controller.ts
@@ -1,8 +1,8 @@
 import * as courseService from "./course.service";
+import * as studentService from "../student/student.service";
 import { createResponse } from "../common/helper/response.hepler";
 import asyncHandler from "express-async-handler";
 import { type Request, type Response } from "express";
-import { getPaginationOptions } from "../common/helper/util.helper";
 import { CourseStatus } from "./course.dto";
 
 export const createCourse = asyncHandler(
@@ -16,7 +16,7 @@ export const updateCourse = asyncHandler(
   async (req: Request, res: Response) => {
     const course = await courseService.getCourseById(req.params.id);
     if (!course) {
-      throw new Error("Course not found")
+      throw new Error("Course not found");
     }
     const updateBody = req.body;
     if (updateBody.duration != course.semesters.length) {
@@ -32,7 +32,7 @@ export const updateCourse = asyncHandler(
 export const editCourse = asyncHandler(async (req: Request, res: Response) => {
   const course = await courseService.getCourseById(req.params.id);
   if (!course) {
-    throw new Error("Course not found")
+    throw new Error("Course not found");
   }
   const editBody = req.body;
   if (editBody.duration) {
@@ -49,6 +49,10 @@ export const editCourse = asyncHandler(async (req: Request, res: Response) => {
 
 export const deleteCourse = asyncHandler(
   async (req: Request, res: Response) => {
+    const enrolledStudentsCount = await studentService.getCourseStudentCount(req.params.id);
+    if(enrolledStudentsCount > 0) {
+      throw new Error("Cannot delete course with enrolled students")
+    }
     const result = await courseService.deleteCourse(req.params.id);
     res.send(createResponse(result, "Course deleted sucssefully"));
   },
@@ -64,7 +68,9 @@ export const getCourseById = asyncHandler(
 export const getAllCourse = asyncHandler(
   async (req: Request, res: Response) => {
     const status = req.query.status as CourseStatus;
-    const isPopulateSemsters = Boolean(parseInt(req.query.isPopulateSemsters as string))
+    const isPopulateSemsters = Boolean(
+      parseInt(req.query.isPopulateSemsters as string),
+    );
     const result = await courseService.getAllCourse(status, isPopulateSemsters);
     res.send(createResponse(result));
   },
